add Selection of run iPerf3 instances and upload them manually

See merge request ngni/opennetworktoolkit/OpenMobileNetworkToolkit!18

+++ app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/Iperf3/JSON/Udp.java
@@ -18,7 +18,7 @@ public class Udp{
     public Float seconds;
     @SerializedName("bytes")
     @Expose
-    public Integer bytes;
+    public Long bytes;
     @SerializedName("bits_per_second")
     @Expose
     public Float bitsPerSecond;
